Refatora a view das apresentações
