+++ resources/views/talks/intro.blade.php
@@ -1,0 +1,7 @@
+<section>
+	<h2>5º Seminário Locaweb PHPSP!</h2>
+	<h3>Smartphone / Tablet / Notebook</h3>
+	<p><img src="images/navegadores.png" /></p>
+	<p><a href="http://seminariophpsp.jelasticlw.com.br/">http://seminariophpsp.jelasticlw.com.br/</a></p>
+	<h4><a href="https://github.com/PHPSP/docs/blob/master/codigo-de-conduta.md">Código de Conduta</a></h4>
+</section>
