+++ resources/views/talks/talk1.blade.php
@@ -1,0 +1,497 @@
+<!-- slide-01 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-01.jpeg">
+</section>
+
+<!-- slide-02 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-02.jpeg">
+</section>
+
+<!-- slide-03 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-03.jpeg">
+</section>
+
+<!-- slide-04 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-04.jpeg">
+</section>
+
+<!-- slide-05 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-05.jpeg">
+</section>
+
+<!-- pool-01 -->
+<section data-background="#424242" class="poll functional-poll" data-number="funcional-1">
+    <h1 class="poll-title">Pergunta!</h1>
+    <p class="poll-question">Você sabe o que é programação funcional?</p>
+
+    <div class="button-level" data-value="sim">
+        <span>Sim (<b>0</b>)</span>
+        <div class="level green"></div>
+    </div>
+
+    <div class="button-level" data-value="não">
+        <span>Não (<b>0</b>)</span>
+        <div class="level red"></div>
+    </div>
+</section>
+
+<!-- slide-06 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-06.jpeg">
+</section>
+
+<!-- slide-07 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-07.jpeg">
+</section>
+
+<!-- slide-08 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-08.jpeg">
+</section>
+
+<!-- slide-09 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-09.jpeg">
+</section>
+
+<!-- slide-10 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-10.jpeg">
+</section>
+
+<!-- slide-11 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-11.jpeg">
+</section>
+
+<!-- slide-12 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-12.jpeg">
+</section>
+
+<!-- slide-13 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-13.jpeg">
+</section>
+
+<!-- slide-14 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-14.jpeg">
+</section>
+
+<!-- slide-15 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-15.jpeg">
+</section>
+
+<!-- slide-16 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-16.jpeg">
+</section>
+
+<!-- slide-17 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-17.jpeg">
+</section>
+
+<!-- slide-18 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-18.jpeg">
+</section>
+
+<!-- slide-19 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-19.jpeg">
+</section>
+
+<!-- slide-20 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-20.jpeg">
+</section>
+
+<!-- slide-21 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-21.jpeg">
+</section>
+
+<!-- slide-22 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-22.jpeg">
+</section>
+
+<!-- slide-23 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-23.jpeg">
+</section>
+
+<!-- slide-24 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-24.jpeg">
+</section>
+
+<!-- slide-25 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-25.jpeg">
+</section>
+
+<!-- slide-26 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-26.jpeg">
+</section>
+
+<!-- slide-27 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-27.jpeg">
+</section>
+
+<!-- slide-28 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-28.jpeg">
+</section>
+
+<!-- slide-29 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-19.jpeg">
+</section>
+
+<!-- slide-30 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-30.jpeg">
+</section>
+
+<!-- slide-31 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-31.jpeg">
+</section>
+
+<!-- pool-02 -->
+<section data-background="#424242" class="poll functional-poll" data-number="2">
+    <h1 class="poll-title">Pergunta!</h1>
+    <p class="poll-question">A função do slide anterior é uma função pura?</p>
+
+    <div class="button-level" data-value="sim">
+        <span>Sim (<b>0</b>)</span>
+        <div class="level green"></div>
+    </div>
+
+    <div class="button-level" data-value="não">
+        <span>Não (<b>0</b>)</span>
+        <div class="level red"></div>
+    </div>
+</section>
+
+<!-- slide-32 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-32.jpeg">
+</section>
+
+<!-- slide-33 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-33.jpeg">
+</section>
+
+<!-- slide-34 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-34.jpeg">
+</section>
+
+<!-- slide-35 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-35.jpeg">
+</section>
+
+<!-- slide-36 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-36.jpeg">
+</section>
+
+<!-- slide-37 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-37.jpeg">
+</section>
+
+<!-- slide-38 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-38.jpeg">
+</section>
+
+<!-- slide-39 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-39.jpeg">
+</section>
+
+<!-- slide-40 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-40.jpeg">
+</section>
+
+<!-- slide-41 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-41.jpeg">
+</section>
+
+<!-- slide-42 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-42.jpeg">
+</section>
+
+<!-- slide-43 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-43.jpeg">
+</section>
+
+<!-- slide-44 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-44.jpeg">
+</section>
+
+<!-- slide-45 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-45.jpeg">
+</section>
+
+<!-- pool-03 -->
+<section data-background="#424242" class="poll functional-poll" data-number="3">
+    <h1 class="poll-title">Pergunta!</h1>
+    <p class="poll-question">Qual será o resultado de <code>print $sayHello('John')</code> do slide anterior?</p>
+
+    <div class="button-level" data-value="hello-john">
+        <span>Hello John! (<b>0</b>)</span>
+        <div class="level green"></div>
+    </div>
+
+    <div class="button-level" data-value="hello">
+        <span>Hello ! (<b>0</b>)</span>
+        <div class="level red"></div>
+    </div>
+</section>
+
+<!-- slide-46 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-46.jpeg">
+</section>
+
+<!-- slide-47 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-47.jpeg">
+</section>
+
+<!-- slide-48 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-48.jpeg">
+</section>
+
+<!-- slide-49 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-49.jpeg">
+</section>
+
+<!-- slide-50 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-50.jpeg">
+</section>
+
+<!-- slide-51 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-51.jpeg">
+</section>
+
+<!-- slide-52 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-52.jpeg">
+</section>
+
+<!-- slide-53 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-53.jpeg">
+</section>
+
+<!-- slide-54 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-54.jpeg">
+</section>
+
+<!-- slide-55 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-55.jpeg">
+</section>
+
+<!-- slide-56 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-56.jpeg">
+</section>
+
+<!-- slide-57 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-57.jpeg">
+</section>
+
+<!-- slide-58 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-58.jpeg">
+</section>
+
+<!-- slide-59 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-59.jpeg">
+</section>
+
+<!-- slide-60 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-60.jpeg">
+</section>
+
+<!-- slide-61 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-61.jpeg">
+</section>
+
+<!-- slide-62 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-62.jpeg">
+</section>
+
+<!-- slide-63 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-63.jpeg">
+</section>
+
+<!-- slide-64 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-64.jpeg">
+</section>
+
+<!-- slide-65 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-65.jpeg">
+</section>
+
+<!-- slide-66 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-66.jpeg">
+</section>
+
+<!-- slide-67 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-67.jpeg">
+</section>
+
+<!-- slide-68 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-68.jpeg">
+</section>
+
+<!-- slide-69 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-69.jpeg">
+</section>
+
+<!-- slide-70 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-70.jpeg">
+</section>
+
+<!-- slide-71 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-71.jpeg">
+</section>
+
+<!-- slide-72 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-72.jpeg">
+</section>
+
+<!-- slide-73 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-73.jpeg">
+</section>
+
+<!-- slide-74 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-74.jpeg">
+</section>
+
+<!-- slide-75 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-75.jpeg">
+</section>
+
+<!-- slide-76 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-76.jpeg">
+</section>
+
+<!-- slide-77 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-77.jpeg">
+</section>
+
+<!-- slide-78 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-78.jpeg">
+</section>
+
+<!-- slide-79 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-79.jpeg">
+</section>
+
+<!-- slide-80 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-80.jpeg">
+</section>
+
+<!-- slide-81 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-81.jpeg">
+</section>
+
+<!-- slide-82 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-82.jpeg">
+</section>
+
+<!-- slide-83 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-83.jpeg">
+</section>
+
+<!-- slide-84 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-84.jpeg">
+</section>
+
+<!-- slide-85 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-85.jpeg">
+</section>
+
+<!-- slide-86 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-86.jpeg">
+</section>
+
+<!-- slide-87 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-87.jpeg">
+</section>
+
+<!-- slide-88 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-88.jpeg">
+</section>
+
+<!-- slide-89 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-89.jpeg">
+</section>
+
+<!-- slide-90 -->
+<section data-background="#424242">
+    <img src="images/funcional/slide-90.jpeg">
+</section>
